requirements.txt
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
pytest>=7.0
openai>=1.0.0
python-dotenv>=1.0.0

# Stage 2: Admin Agent & Telegram Integration
python-telegram-bot>=20.0
requests>=2.31.0
langchain>=0.1.0
